exfat: fix uninit-value in __exfat_get_dentry_set
There is no check if stream size and start_clu are invalid.
If start_clu is EOF cluster and stream size is 4096, It will cause
uninit value access. because ei->hint_femp.eidx could be 128(if cluster
size is 4K) and wrong hint will allocate next cluster. and this cluster
will be same with the cluster that is allocated by
exfat_extend_valid_size(). The previous patch will check invalid start_clu,
but for clarity, Initialize hint_femp.eidx to zero.

Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>